Export and reuse TestDocumentBuilderSupplier
